loadingScene : initialization in loading scene
